Revert "getmemsize back to size_t"
This reverts commit 35b19b9a89e63b6405dbdfa3c056fe095530d4f7.

+++ Buffer_shared.cpp
@@ -77,7 +77,7 @@ void FacetHistogramBuffer::Reset() {
 #endif
 }
 
-size_t FacetHistogramBuffer::GetMemSize() const {
+int FacetHistogramBuffer::GetMemSize() const {
 	int sum = 0;
 	sum += nbHitsHistogram.capacity() * sizeof(double);
 	sum += distanceHistogram.capacity() * sizeof(double);
